+++ amd/src/editor.js
@@ -69,14 +69,17 @@ define(['jquery'], function($) {
                     }
 
                     if (enableapptoggle[0].checked
-                        && formatelement.length > 0 && formatelement[0].value == 'docxf') {
+                        && formatelement.length > 0
+                        && formatelement[0].value == 'pdf') {
                         openEditor(contextid, itemid, readonly, tmplkey);
                     }
 
                     enableapptoggle.change(function(e) {
                         if (e.currentTarget.checked
-                            && (formatelement.length > 0 && formatelement[0].value == 'docxf'
-                            || selectformat.length > 0 && selectformat[0].value == 'docxf')) {
+                            && (formatelement.length > 0
+                                && formatelement[0].value == 'pdf'
+                                || selectformat.length > 0
+                                && selectformat[0].value == 'pdf')) {
 
                             if (docEditor === null) {
                                 openEditor(contextid, itemid, readonly, tmplkey);
@@ -92,7 +95,7 @@ define(['jquery'], function($) {
 
                     if (selectformat.length > 0) {
                         selectformat.change(function(e) {
-                            if (e.target.value != 'docxf') {
+                            if (e.target.value != 'pdf') {
                                 if ($("#app-onlyoffice").is(":visible")) {
                                     $("#app-onlyoffice").hide();
                                 }