+++ lang/de/assignsubmission_onlyoffice.php
@@ -23,10 +23,10 @@
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 $string['assignformat'] = 'Format (ONLYOFFICE)';
-$string['docxfformname'] = 'Formular';
 $string['enabled'] = 'ONLYOFFICE-Dokument';
 $string['enabled_help'] = 'Wenn diese Option aktiviert ist, können die Studierenden ONLYOFFICE für ihre Einreichung verwenden.';
 $string['filenotfound'] = 'Datei nicht gefunden';
 $string['formnotready'] = 'Das Formular ist noch nicht fertig. Bitte warten Sie ein wenig und laden Sie die Seite neu.';
+$string['pdfformname'] = 'Formular';
 $string['pluginname'] = 'ONLYOFFICE-Dokument';
 $string['viewdocument'] = 'ONLYOFFICE-Dokument anzeigen';